feat: Parallel Compact

+++ lsm/cache.go
@@ -35,6 +35,6 @@ func newCache(opt *Options) *cache {
 }
 
 // TODO fid 使用字符串是不是会有性能损耗
-func (c *cache) addIndex(fid uint32, t *table) {
+func (c *cache) addIndex(fid uint64, t *table) {
 	c.indexs.Set(fid, t)
 }
